+++ test/compojure/api/sweet_test.clj
@@ -18,7 +18,8 @@
 
 (def app
   (api
-    {:swagger {:spec "/swagger.json"
+    {:formatter :muuntaja
+     :swagger {:spec "/swagger.json"
                :data {:info {:version "1.0.0"
                              :title "Sausages"
                              :description "Sausage description"
@@ -210,7 +211,8 @@
 
 (deftest produces-and-consumes-test
   (let [app (api
-              {:swagger {:spec "/swagger.json"
+              {:formatter :muuntaja
+               :swagger {:spec "/swagger.json"
                          :data {:produces ["application/json" "application/edn"]
                                 :consumes ["application/json" "application/edn"]}}}
               ping-route)]